Improved KingpinStateView
