+++ GACore.Controls/GACore.Controls.csproj
@@ -61,18 +61,9 @@
     <Compile Include="IsConnectedControl.xaml.cs">
       <DependentUpon>IsConnectedControl.xaml</DependentUpon>
     </Compile>
-    <Compile Include="KingpinStateControl.xaml.cs">
-      <DependentUpon>KingpinStateControl.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="KingpinStatusControl.xaml.cs">
-      <DependentUpon>KingpinStatusControl.xaml</DependentUpon>
-    </Compile>
     <Compile Include="LabelContentControl.cs" />
     <Compile Include="LightStateControl.xaml.cs">
       <DependentUpon>LightStateControl.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="PoseControl.xaml.cs">
-      <DependentUpon>PoseControl.xaml</DependentUpon>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs">
       <SubType>Code</SubType>
@@ -96,6 +87,9 @@
     </Compile>
     <Compile Include="View\KingpinStatusView.xaml.cs">
       <DependentUpon>KingpinStatusView.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="View\PoseView.xaml.cs">
+      <DependentUpon>PoseView.xaml</DependentUpon>
     </Compile>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
@@ -130,19 +124,7 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="KingpinStateControl.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="KingpinStatusControl.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
     <Page Include="LightStateControl.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="PoseControl.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -155,6 +137,10 @@
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="View\KingpinStatusView.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="View\PoseView.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
