--- GACore.Controls/KingpinStateControl.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace GACore.Controls
-{
-	/// <summary>
-	/// Interaction logic for KingpinStateControl.xaml
-	/// </summary>
-	public partial class KingpinStateControl : UserControl
-	{
-		public KingpinStateControl()
-		{
-			InitializeComponent();
-		}
-	}
-}
