--- GACore.Controls/KingpinStatusControl.xaml
@@ -1,62 +1,0 @@
-﻿<UserControl x:Class="GACore.Controls.KingpinStatusControl"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:local="clr-namespace:GACore.Controls"
-             mc:Ignorable="d"
-             d:DesignHeight="450" d:DesignWidth="800">
-    <UserControl.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="pack://application:,,,/GACore.Controls;component/ResourceDictionary.xaml" />
-            </ResourceDictionary.MergedDictionaries>
-            <local:PositionControlStatusToOverlayTextConverter x:Key="pcsConverter" />
-            <local:NavigationStatusToOverlayTextConverter x:Key="nsConverter" />
-            <local:DynamicLimiterStatusToOverlayTextConverter x:Key="dlsConverter" />
-        </ResourceDictionary>
-    </UserControl.Resources>
-    <Border BorderBrush="Black">
-        <StackPanel>
-            <Grid IsHitTestVisible="True" Width="150" Height="30" ToolTip="Status of the vehicle's controller system"
-                  Background="{Binding Path=PositionControlStatus, Converter={StaticResource pcsConverter}, ConverterParameter={x:Static local:BrushCollectionProperty.Background}}" Opacity="0.5">
-                <Viewbox HorizontalAlignment="Stretch" Panel.ZIndex="2">
-                    <local:LabelContentControl Header="Pos:" Foreground="{Binding Path=PositionControlStatus, Converter={StaticResource pcsConverter},
-                        ConverterParameter={x:Static local:BrushCollectionProperty.Foreground}, Mode=OneWay}">
-                        <TextBlock HorizontalAlignment="Center" VerticalAlignment="Center"
-                               Text="{Binding Path=PositionControlStatus,
-                               Converter={StaticResource pcsConverter}, ConverterParameter={x:Static local:BrushCollectionProperty.Text}, FallbackValue=DataContext?}"
-                               Foreground="{Binding Path=PositionControlStatus, Converter={StaticResource pcsConverter}, ConverterParameter={x:Static local:BrushCollectionProperty.Foreground}}">
-                        </TextBlock>
-                    </local:LabelContentControl>
-                </Viewbox>
-            </Grid>
-            <Separator Margin="0" />
-            <Grid IsHitTestVisible="True" Width="150" Height="30" ToolTip="Status of the vehicle's navigation system"
-                  Background="{Binding Path=NavigationStatus, Converter={StaticResource nsConverter}, ConverterParameter={x:Static local:BrushCollectionProperty.Background}}" Opacity="0.5">
-                <Viewbox HorizontalAlignment="Stretch" Panel.ZIndex="2">
-                    <local:LabelContentControl Header="Nav:" Foreground="{Binding Path=NavigationStatus, Converter={StaticResource nsConverter}, ConverterParameter={x:Static local:BrushCollectionProperty.Foreground}, Mode=OneWay}">
-                        <TextBlock HorizontalAlignment="Center" VerticalAlignment="Center"
-                                   Text="{Binding Path=NavigationStatus,
-                                   Converter={StaticResource nsConverter}, ConverterParameter={x:Static local:BrushCollectionProperty.Text}, FallbackValue=DataContext?}"
-                                   Foreground="{Binding Path=NavigationStatus, Converter={StaticResource nsConverter}, ConverterParameter={x:Static local:BrushCollectionProperty.Foreground}}">
-                        </TextBlock>
-                    </local:LabelContentControl>
-                </Viewbox>
-            </Grid>
-            <Separator Margin="0" />
-            <Grid IsHitTestVisible="True" Width="150" Height="30" ToolTip="Status of the vehicle's safety systems"
-                Background="{Binding Path=DynamicLimiterStatus, Converter={StaticResource dlsConverter}, ConverterParameter={x:Static local:BrushCollectionProperty.Background}}" Opacity="0.5">
-                <Viewbox HorizontalAlignment="Stretch" Panel.ZIndex="2">
-                    <local:LabelContentControl Header="Lim:" Foreground="{Binding Path=DynamicLimiterStatus, Converter={StaticResource dlsConverter}, ConverterParameter={x:Static local:BrushCollectionProperty.Foreground}, Mode=OneWay}">
-                        <TextBlock HorizontalAlignment="Center" VerticalAlignment="Center"
-                                Text="{Binding Path=DynamicLimiterStatus,
-                                Converter={StaticResource dlsConverter}, ConverterParameter={x:Static local:BrushCollectionProperty.Text}, FallbackValue=DataContext?}"
-                                Foreground="{Binding Path=DynamicLimiterStatus, Converter={StaticResource dlsConverter}, ConverterParameter={x:Static local:BrushCollectionProperty.Foreground}}">
-                        </TextBlock>
-                    </local:LabelContentControl>
-                </Viewbox>
-            </Grid>
-        </StackPanel>
-    </Border>
-</UserControl>
