--- GACore.Controls/KingpinStatusControl.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace GACore.Controls
-{
-	/// <summary>
-	/// Interaction logic for KingpinStatusControl.xaml
-	/// </summary>
-	public partial class KingpinStatusControl : UserControl
-	{
-		public KingpinStatusControl()
-		{
-			InitializeComponent();
-		}
-	}
-}
