--- GACore.Controls/PoseControl.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace GACore.Controls
-{
-	/// <summary>
-	/// Interaction logic for PoseControl.xaml
-	/// </summary>
-	public partial class PoseControl : UserControl
-	{
-		public PoseControl()
-		{
-			InitializeComponent();
-		}
-	}
-}
