+++ GACore.DemoApp/MainWindow.xaml
@@ -22,7 +22,6 @@
         <StackPanel>
             <gaCoreUC:IPV4Control IPV4String="192.168.1.1" />
             <view:KingpinStateView x:Name="kingpinStateView"/>
-            <gaCoreUC:KingpinStateControl x:Name="kingpinStateControl" DataContext="{Binding Source={StaticResource fooKingpin}, Path=KingpinState}" />
             <gaCoreUC:LabelContentControl Header="Good">
                 <Button x:Name="goodButton" Click="GoodButton_Click">Good</Button>
             </gaCoreUC:LabelContentControl>
